feat: support boolean values encoded as text
status-go, owing to its use of a version of sqlcipher that uses a version of
sqlite previous to 3.23.0, sometimes inserts (“0”, 0, false, FALSE, “false”,
“FALSE”) or (“1”, 1, true, TRUE, “true”, “TRUE”) for boolean values in the
database. Here, we are able to properly decode text and integer encodings of
booleans in the database into Nim's `bool` type.

feat: add unit tests for sqlite boolean literals support